Apply QuestionCallback to the Indicators
